Added notifications list component for separation of concerns
